Small fixes

+++ db/migrate/20190528165800_create_ski_routes.rb
@@ -3,9 +3,9 @@ class CreateSkiRoutes < ActiveRecord::Migration[6.0]
     create_table :ski_routes do |t|
       t.string :title
       t.text :intro
-      t.integer :distance
+      t.decimal :distance
       t.integer :vert
-      t.integer :time
+      t.decimal :time
       t.string :aspect
       t.integer :dificulty
       t.text :route_description
